moved styled-components for modals into modal directory
