+++ src/components/Helpers/Modal/CloseModalButtonStyle.js
