--- src/styles/Helpers/Modals/index.js
@@ -1,11 +1,0 @@
-import ModalContainer from "./ModalContainer";
-import MatchModal from "./MatchModal";
-import CloseModalButton from "./CloseModalButton";
-
-const ModalStyle = {
-  ModalContainer,
-  MatchModal,
-  CloseModalButton
-};
-
-export default ModalStyle;
