+++ src/styles/Helpers/index.js
@@ -1,6 +1,6 @@
 import { LikedHeart, NotLikedHeart } from "./LikedHeart";
 import UserLike from "./UserLike";
-import ModalStyle from "./Modals";
+
 import ImageWrapper from "./ImageWrapperStyle";
 
 export const ImageWrapperStyle = ImageWrapper;
@@ -9,10 +9,3 @@ export const ImageWrapperStyle = ImageWrapper;
 export const UserLikeStyle = UserLike;
 export const LikedHeartStyle = LikedHeart;
 export const NotLikedHeartStyle = NotLikedHeart;
-
-// ModalStyle components
-export const ModalContainerStyle = ModalStyle.ModalContainer;
-
-export const CloseModalButtonStyle = ModalStyle.CloseModalButton;
-
-export const MatchModalStyle = ModalStyle.MatchModal;
